Move sources_page.dart to sources/page file
